+++ packages/terra-clinical-item-display/src/ItemComment.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import CommentIndicator from 'terra-icon/lib/icon/IconComment';
-/* eslint-disable-next-line import/no-cycle */
 import ItemDisplay from './ItemDisplay';
 import styles from './ItemComment.scss';