+++ packages/terra-clinical-item-display/src/ItemDisplay.jsx
@@ -2,8 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import styles from './ItemDisplay.scss';
-/* eslint-disable-next-line import/no-cycle */
-import ItemComment from './ItemComment';
 
 const cx = classNames.bind(styles);
 
@@ -91,8 +89,6 @@ const ItemDisplay = ({
 
 ItemDisplay.propTypes = propTypes;
 ItemDisplay.defaultProps = defaultProps;
-
-ItemDisplay.Comment = ItemComment;
 
 export default ItemDisplay;
 export { TextStyles };