+++ packages/terra-clinical-item-display/src/terra-dev-site/doc/example/DefaultComment.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
-import ItemDisplay from 'terra-clinical-item-display/lib/ItemDisplay';
+import ItemComment from 'terra-clinical-item-display/lib/ItemComment';
 
 const component = () => (
   <span>
-    <ItemDisplay.Comment text="Comment" />
-    <ItemDisplay.Comment text="Disabled Comment" isDisabled />
+    <ItemComment text="Comment" />
+    <ItemComment text="Disabled Comment" isDisabled />
   </span>
 );