+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Resolved dependency cycle.
+
 ### Changed
 * Add comments for ESlint 6